V0.1.6 #1
Moderator component was added since the database roles are finally implemented. The overall html and logic have to be migrated from the admin component. The current admin component will be replaced with a new one that is able to manage and create new tables on the database + users for the departments and optionally also view all available plans